requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.4.0
aiosqlite==0.20.0
alerts-in-ua==0.3.2
APScheduler==3.11.0
async-timeout==4.0.3
frozenlist==1.7.0
idna==3.3
magic-filter==1.0.12
multidict==6.6.4
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.1
python-telegram-bot==22.3
typing_extensions==4.15.0
tzlocal==5.3.1
yarl==1.20.1
environs
deep-translator
